+++ src/routes/(admin)/admin/eventsadmin/[eventId]/+page.svelte
@@ -2,8 +2,11 @@
 	import { updateDoc } from 'firebase/firestore';
 	import { DocumentReference, type DocumentData } from 'firebase/firestore';
 	import { type DomainEvent, EditModeStore } from '$lib/stores/ObjectStore';
+	import { notificationStore } from '$lib/stores/notifications';
 
 	import EventForm from '$lib/components/EventForm.svelte';
+	import ToastContainer from '$lib/components/ToastContainer.svelte';
+	import { goto } from '$app/navigation';
 
 	type Params = {
 		newEvent: DomainEvent;
@@ -30,12 +33,19 @@
 			const eventData = { ...updatedEvent } as DocumentData;
 			await updateDoc(data.docRef, eventData);
 			EditModeStore.set('');
+			notificationStore.addToast('success', 'Event updated successfully!');
+			goto('/admin/eventsadmin');
 		} catch (error) {
+    	notificationStore.addToast('error', 'Failed to update event. Please try again.');
 			console.error('Error updating the event: ', error);
 		}
 	};
+
 </script>
 
 <div>
 	<EventForm thisEvent={data.newEvent} onUpdate={updateEvent} />
 </div>
+
+<ToastContainer />
+