+++ changelog.md
@@ -12,7 +12,9 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Added
 
--
+- add notifications to provide transaction feedback to the user
+- added notificationStore, ToastContainer and ToastItem components
+- added toast notification handling for event update page
 
 ### Changed
 
@@ -28,11 +30,13 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 - major update to event and news page to use UTC dates instead of local dates
 - implemented a fall back for older items without UTC dates
+
 ## [1.2.0] - 2025-06-24
 
 ### Fixed
 
 - the alt property of images on the events and news detail page are now set to the imageAlt property of the event or news item
+
 ## [1.1.0] - 2025-06-24
 
 ### Added
@@ -49,6 +53,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 - FutureEvents are all events with PublishEndDate > today
 - Online events only show the name of the online service but no link
 - Event metadata on the event detail page is now fully responsive
+
 ## [1.0.0] - 2025-06-24
 
 ### Changed
@@ -64,6 +69,7 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 ### Fixed
 
 - location form didn't come up when called from event admin form
+
 ## [1.0.0] - 2025-06-24
 
 ### Added